requirements.txt
cffi==1.15.0
msgpack==1.0.3
py-algorand-sdk==1.9.0
pycparser==2.21
pycryptodomex==3.14.0
PyNaCl==1.5.0
python-dotenv~=0.19.2
pyteal==0.9.1